+++ src/components/BlockHeightButton/BlockHeightButton.tsx
@@ -47,10 +47,11 @@ const BlockHeightButton: React.FC = () => {
 
   return (
     <Tooltip title="Block height">
-      <Button size="medium" variant="contained">
-        <WidgetsIcon sx={{ fontSize: "14px" }} />
-        {BlockNum}
-
+      <Button variant="contained">
+        <Box sx={{ "@media (max-width: 350px)": { display: "none" } }}>
+          <WidgetsIcon sx={{ fontSize: "14px" }} />
+          {BlockNum}
+        </Box>
         <CircularProgressWithLabel
           value={blockTimeCounter}
           sx={{ color: blockTimeCounter > 22 ? (blockTimeCounter > 66 ? "red" : "blue") : "black" }}