+++ RELEASE-NOTES.md
@@ -63,8 +63,6 @@
 - added tooltip to tier label
 - adjusted dao info page text (summary)
 
-- add farm tier label to the Tentacle farms (tier 1, tier 2, etc...)
-
 DAO Info Page
 - ecosystemTvl added
 - circulating INK added NOTE: Calculates based on harvested INK. IE, unharvested INK is not circulating based on this calculation.
@@ -76,4 +74,8 @@ DAO Info Page
 - ToS added to footer
 
 ## Unreleased
-nothing at this time
+- build version shown in footer
+- links to charts added
+- mobile style updates
+- Harvest all button to trigger a harvest for each farm with an unharvested balance
+- APY is now 0% when contract `isPaused` or current date is after `periodFinish`